Move lazy-load-script and preload-script to base-utils
