+++ assets/js/base/utils/preload-script.ts
